requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
paho-mqtt==1.6.1
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
joblib==1.3.2
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
python-socketio==5.10.0
